[#40] refactor: CommentTest 리팩토링
