+++ src/test/java/com/mallang/comment/domain/service/CommentDeleteServiceTest.java
@@ -1,12 +1,14 @@
 package com.mallang.comment.domain.service;
 
 import static com.mallang.member.MemberFixture.memberBuilder;
+import static com.mallang.member.MemberFixture.말랑;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import com.mallang.comment.domain.AuthenticatedComment;
 import com.mallang.comment.domain.CommentRepository;
+import com.mallang.comment.domain.UnAuthenticatedComment;
 import com.mallang.member.domain.Member;
 import com.mallang.post.domain.Post;
 import org.junit.jupiter.api.DisplayName;
@@ -24,11 +26,11 @@ class CommentDeleteServiceTest {
 
     private final Member postWriter = memberBuilder().id(1L).build();
     private final Post post = Post.builder().member(postWriter).build();
+    private final Member member = 말랑(1L);
 
     @Test
     void 자식_댓글이_존재한다면_부모와의_연관관계는_유지되며_논리적으로만_제거시킨다() {
         // given
-        Member member = memberBuilder().id(1L).build();
         AuthenticatedComment parentComment = AuthenticatedComment.builder()
                 .content("내용")
                 .post(post)
@@ -54,25 +56,17 @@ class CommentDeleteServiceTest {
     @Test
     void 대댓글을_삭제하는_경우_물리적으로_제거된다() {
         // given
-        Member member = memberBuilder().id(1L).build();
-        AuthenticatedComment parentComment = AuthenticatedComment.builder()
+        UnAuthenticatedComment parentComment = UnAuthenticatedComment.builder()
                 .content("내용")
                 .post(post)
-                .writer(member)
-                .secret(true)
+                .nickname("익")
+                .password("1234")
                 .build();
-        AuthenticatedComment childComment = AuthenticatedComment.builder()
+        UnAuthenticatedComment childComment = UnAuthenticatedComment.builder()
                 .content("to be delete")
                 .post(post)
-                .writer(member)
-                .secret(true)
-                .parent(parentComment)
-                .build();
-        AuthenticatedComment childComment2 = AuthenticatedComment.builder()
-                .content("내용")
-                .post(post)
-                .writer(member)
-                .secret(true)
+                .nickname("익")
+                .password("1234")
                 .parent(parentComment)
                 .build();
 
@@ -87,12 +81,11 @@ class CommentDeleteServiceTest {
     @Test
     void 대댓글을_삭제하는_경우_부모_댓글이_논리적으로_제거된_상태이며_더이상_존재하는_자식이_없는_경우_부모_댓글도_물리적으로_제거된다() {
         // given
-        Member member = memberBuilder().id(1L).build();
-        AuthenticatedComment parentComment = AuthenticatedComment.builder()
+        UnAuthenticatedComment parentComment = UnAuthenticatedComment.builder()
                 .content("내용")
                 .post(post)
-                .writer(member)
-                .secret(true)
+                .nickname("익")
+                .password("1234")
                 .build();
         AuthenticatedComment childComment = AuthenticatedComment.builder()
                 .content("to be delete")
@@ -101,7 +94,7 @@ class CommentDeleteServiceTest {
                 .secret(true)
                 .parent(parentComment)
                 .build();
-        parentComment.delete(member, commentDeleteService);
+        parentComment.delete(member, "1234", commentDeleteService);
 
         // when
         commentDeleteService.delete(childComment);
@@ -114,18 +107,17 @@ class CommentDeleteServiceTest {
     @Test
     void 대댓글을_삭제하는_경우_부모_댓글이_논리적으로_제거된_상태가_아닌_경우_부모는_변함없다() {
         // given
-        Member member = memberBuilder().id(1L).build();
         AuthenticatedComment parentComment = AuthenticatedComment.builder()
                 .content("내용")
                 .post(post)
                 .writer(member)
                 .secret(true)
                 .build();
-        AuthenticatedComment childComment = AuthenticatedComment.builder()
+        UnAuthenticatedComment childComment = UnAuthenticatedComment.builder()
                 .content("to be delete")
                 .post(post)
-                .writer(member)
-                .secret(true)
+                .nickname("익")
+                .password("1234")
                 .parent(parentComment)
                 .build();
